+++ types/adapter/entry-points/console/consoleGithubTokenResolver.d.ts
@@ -4,5 +4,6 @@ export declare const extractRepositoryOwner: (issueOrPullRequestUrl: string) => 
 export declare const createConsoleIssueRepositoryResolver: <IssueRepositoryType>(resolveGithubToken: ConsoleGithubTokenResolver, buildIssueRepositoryForToken: (githubToken: string) => IssueRepositoryType) => ((issueOrPullRequestUrl: string) => IssueRepositoryType);
 export declare const extractProjectOwner: (projectUrl: string) => string | null;
 export declare const createConsoleProjectRepositoryResolver: <ProjectRepositoryType>(resolveGithubToken: ConsoleGithubTokenResolver, buildProjectRepositoryForToken: (githubToken: string) => ProjectRepositoryType) => ((projectUrl: string) => ProjectRepositoryType);
+export declare const createConsoleGithubTokenResolverByItemUrl: (resolveGithubToken: ConsoleGithubTokenResolver) => ((itemUrl: string) => string);
 export declare const createConsoleGithubTokenResolver: (defaultToken: string, githubTokenFilePathByRepositoryOwner: Record<string, string> | null, readTokenFile: GithubTokenFileReader) => ConsoleGithubTokenResolver;
 //# sourceMappingURL=consoleGithubTokenResolver.d.ts.map